OPENEUROPA-2668: Create DateFieldContext.
